+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 * **terra-slider** add focus highlighting for tab navigation
+* **terra-select-box** added key events for better handling
 
 <a name="3.0.0-beta.18"></a>
 # 3.0.0-beta.18 (06.12.2018)